adding auto-expansion for heatmap if datatype is configured to expand
made columndefs adjust dynamically to the number of cols

this is necessary for when some datatypes are configured to
show up as "expanded"

flexible sorting setup: takes a variable set of datatypes + datasources

this is necessary for when datatypes are configured to
show as "expanded". This code also removed some hard coded
references to datatypes.

solution for displaying custom datasource panel, if one is configured

Normally datasources are displayed together in the
datatype pannel. This option allows for a datasource
to be configured to have its own panel.

remove link to to main datatype's panel when config is show as expanded

in some cases, the data source can have a different panel to visualize the data.
This change ensures the links in the main heatmap
association page make sense.

solution for target association page